requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
alembic
antlr4-python3-runtime
asteroid-filterbanks
asttokens
attrs
av
certifi
cffi
charset-normalizer
click
coloredlogs
colorlog
contourpy
ctranslate2
cycler
decorator
docopt
einops
executing
faster-whisper
filelock
flatbuffers
fonttools
frozenlist
fsspec
greenlet
hf-xet
huggingface-hub
humanfriendly
HyperPyYAML
idna
ipdb
ipython
ipython_pygments_lexers
jedi
Jinja2
joblib
julius
kiwisolver
lightning
lightning-utilities
Mako
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mdurl
mpmath
multidict
networkx
numpy
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-cufile-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-cusparselt-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
omegaconf
onnxruntime
optuna
packaging
pandas
parso
pexpect
pillow
primePy
prompt_toolkit
propcache
protobuf
ptyprocess
pure_eval
pyannote.audio
pyannote.core
pyannote.database
pyannote.metrics
pyannote.pipeline
pycparser
Pygments
pyparsing
python-dateutil
pytorch-lightning
pytorch-metric-learning
pytz
PyYAML
requests
rich
ruamel.yaml
ruamel.yaml.clib
scikit-learn
scipy
semver
sentencepiece
setuptools
shellingham
six
sortedcontainers
sounddevice
soundfile
speechbrain
SQLAlchemy
stack-data
sympy
tabulate
tensorboardX
threadpoolctl
tokenizers
torch
torch-audiomentations
torch_pitch_shift
torchaudio
torchmetrics
tqdm
traitlets
triton
typer
typing_extensions
tzdata
urllib3
uv
wcwidth
yarl